26 optimize serialization size (#53)
* Optimize size of Step message by using uint8_t instead of int for wave and step number ==> Save 6 bytes per Step message

* Create adaptCereal.h file in order to define CEREAL_SIZE_TYPE so that Cereal uses only 4 bytes (instead of 8) to store size_t fields.

* Add a return to CommStub::toString() so that code compiles under Windows.

* For Session messages, use Cereal polymorphic capabilities.

---------

+++ src/AlgoLayer/BBOBB/BBOBBMsg.h
@@ -1,6 +1,7 @@
 #ifndef FBAE_BBOBBMSG_H
 #define FBAE_BBOBBMSG_H
 
+#include "../../adaptCereal.h"
 #include "cereal/archives/binary.hpp"
 #include "cereal/types/string.hpp"
 #include "cereal/types/vector.hpp"
@@ -22,8 +23,8 @@ namespace fbae_BBOBBAlgoLayer {
     struct StepMsg {
         MsgId msgId{};
         rank_t senderPos{};
-        int wave;
-        int step;
+        uint8_t wave;
+        uint8_t step;
         std::vector<fbae_AlgoLayer::BatchSessionMsg> batchesBroadcast;
 
         // This method lets cereal know which data members to serialize
